requirements.txt
fastmcp
mysql-connector-python
uvicorn
python-dotenv
